// src/types.ts
export type FocusArea = {
    slug: string;
    title: string;
    description: string;
    primaries: ModuleProps[];
    electives: ModuleProps[];
};

export type ModuleProps = {
    code: string;
    title: string;
};